refactor: place `net.sourceforge.plantuml.test` on `test.utils`
Create `test` folder with:
- `example`
- `utils`

And put all `net.sourceforge.plantuml.test` on `test.utils`.
_[no other change]_
