+++ test/test/utils/TestUtils.java
@@ -1,4 +1,4 @@
-package net.sourceforge.plantuml.test;
+package test.utils;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
